refactor(circle): extract MemberStatusGrid from InCircleView (Sem 5 Día 2)
Moves _MemberListItem + member header/list build section into a pure
MemberStatusGrid widget. InCircleView passes data via callbacks — no
Firestore or static services in the new widget. Also fixes 2
withOpacity → withValues and removes 2 unused _AppTextStyles entries.
